Pass `:broken_record` context to validation call.
The `valid?` method takes an optional context argument. Validations
themselves can take an `on:` kwarg; if they do, they will only run
when the matching symbol is passed to the `valid?` method. Validations
without an `on:` kwarg will run every time.

By passing a `:broken_record` context to `valid?`, we can scope
expensive sanity-check validations to only run during the broken record
pass, and not during normal app operation. After this is deployed, all
that's needed to scope that is to add `on: :broken_record` to that
validation.